+++ libs/react/src/__mocks__/surveyDataMock.ts
@@ -22,11 +22,7 @@ export const getTestingSurveyData = (): SurveyDataType => {
       image_url:
         'https://images.unsplash.com/photo-1603437873662-dc1f44901825?ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D&auto=format&fit=crop&w=3270&q=80',
       intro_markdown:
-        'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.\n' +
-        '\n' +
-        '- Bullet 1\n' +
-        '- Bullet 2\n' +
-        '- Bullet 3',
+        "Welcome to Cold Climate!\n Let’s Start Your Journey to Absolute Zero™ \nWe will start with our basic company information survey. \nThis is a quick form to understand a little more about your company and what climate efforts you've already undertaken.",
       sections: {
         product: {
           title: 'Product',