+++ libs/react/src/components/molecules/footprintDetailCard/footprintDetailCard.stories.tsx
@@ -1,0 +1,48 @@
+import React from 'react';
+import { withKnobs } from '@storybook/addon-knobs';
+import { Meta, StoryObj } from '@storybook/react';
+import { FootprintDetailCard } from './footprintDetailCard';
+import { BrowserRouter } from 'react-router-dom';
+import { getSchemeForColor, HexColors } from '@coldpbc/themes';
+import { render } from 'react-dom';
+
+const meta = {
+  title: 'Molecules/FootprintDetailCard',
+  component: FootprintDetailCard,
+  tags: ['autodocs'],
+  decorators: [withKnobs],
+} satisfies Meta<typeof FootprintDetailCard>;
+
+export default meta;
+type Story = StoryObj<typeof meta>;
+
+export const Default = () => {
+  return (
+    <BrowserRouter>
+      <div className="w-[668px]">
+        <FootprintDetailCard colors={getSchemeForColor(HexColors.lightblue)} period={2022} subcategory_key='facilities' />
+      </div>
+    </BrowserRouter>
+  );
+};
+
+export const GreenProduct = () => {
+  return (
+    <BrowserRouter>
+      <div className="w-[668px]">
+        <FootprintDetailCard colors={getSchemeForColor(HexColors.green)} period={2022} subcategory_key='product' />
+      </div>
+    </BrowserRouter>
+  );
+};
+
+export const EmptySubcategory = () => {
+  return (
+    <BrowserRouter>
+      <div className="w-[668px]">
+        <FootprintDetailCard colors={getSchemeForColor(HexColors.green)} period={2022} subcategory_key='no_data_for_this_key' />
+      </div>
+    </BrowserRouter>
+  );
+};
+